pyproject.toml
[project]
name = "neuro-pilot-labeler"
version = "0.1.0"
description = "An e2e labeler for self driving trajectory prediction and object detection"
readme = "README.md"
requires-python = "==3.10.0"
dependencies = [
    "fastapi>=0.135.1",
    "minio>=7.2.20",
    "opencv-python>=4.13.0.92",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
    "pyyaml",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
